Added the rest of the files for the array-data benchmark to the repository.
load_data.sh generates the data (with the *.c files) and bulk-loads them
into the server (using the schemas in the *.fmt files).

the 3 queries are my best interpretation of what was done in the
RasDaMan performance evaluation paper, some queries may need to
be adjusted (My SQL is a bit rusty...).

finally 'drop.sql' is the basic cleanup file...