The before tag is Before64BitStage3, the after tag is After64BitStage3.
This is step 3 towards the support of 64 bits.

I introduced a new type var_t for the pointers into the heap for
var-sized BATs.  There is also a symbol TYPE_var which gets expanded
to TYPE_int or TYPE_lng, depending on the size of a var_t.
Everywhere where we now point into the heap, we use var_t, and var_t
is 64 bits on a 64-bit machine.

The upshot is: heaps can be bigger than 4 GB on a 64-bit machine.

Lots more testing to do, but things seem to work pretty well.