Fixed mapi_dup/mapi_disconnect interaction:  Mapi structs that share the to/from streams are now linked into a doubly linked list. mapi_dup adds a new entry to the list, mapi_disconnect removes the struct from the list, and if it was the only member of the list, closes the streams.
Split mapi_connect into two functions: mapi_mapi (for lack of a better
name, same args as mapi_connect) and mapi_reconnect.  mapi_mapi
creates a new Mapi struct and fills in all the fields according to the
arguments.  mapi_reconnect connects to the server.  The connection
code that was originally in mapi_connect was moved to mapi_reconnect.
The latter now no longer calls mapi_connect, creating a new Mapi
struct which is then copied to the original one and then destroyed.
The connected field is now only a Boolean value.

mapi_disconnect no longer always returns an error.  Instead, it only
returns an error if the requested operation failed.