API change: mapi_param_type has grown an extra (4th) argument: int sqltype, the type of the SQL value.  Currently we don't actually do anything with this value except store it.
Separated input (parameter) bindings and output bindings.  They're
totally different things.  The former have to do with replacing ? in
SQL commands and the latter with binding output columns to application
variables.

Also made sure that fieldcnt is consistently used for how many columns
(fields) there are in the output table and that maxfields is
consistently used for the allocated space for the column descriptors
(maxfields >= fieldcnt).