The previous checkin was a fix for the problem that a module statement is supposed to make stuff that is defined in the module available for statements on the same line as the module statement.  This was done by returning the string-to-be-executed when a semicolon in seen instead of waiting for the newline.
However, the fix made that module statements were never printed
anymore when the trace flag is on.

This fixes that problem by printing whatever is returned by the parse
function (i.e. everything up to and including the semicolon, and the
next time, everything after the semicolon up to and including the
newline) instead of waiting for a newline and then printing that line.

This fix has, however, one problem: if there is an error in the
statement and tracing is on, the error message appears after the
statement being executed instead of below.  This is because of the
first fix, the command is executed as soon as the semicolon is seen
instead of when the newline is seen, and the second fix prints the
stuff up to and including that semi-colon but not the newline.