disabled my previous patch, again:
It does not work (yet?):
The libreadline in /lib64 seem to have a symbol/function "readline",
but no "rl_history_search_forward", while on all other platform
we seem to have "rl_history_search_forward", but no "readline".
With "readline" instead of "rl_history_search_forward",
64-bit MapiClient makes 64-bit Mserver crash on connect on spin...

Could it be, that the 64-bit libreadline is not properly installed on spin?