Reimplemented, but keeping parts of the original code.
The reason for the change: there are now new functions in Mapi:
mapi_query_prep, mapi_query_part, and mapi_query_done which can be
used to send large queries to the server.  Using these is much, much
more efficient than reallocating memory all the time to send a large
query.  The easiest way to make use of these new functions was to
reimplement since the code was too complex.

What we gained: I find it much more understandable :-) .  There are no
longer macros with side effects.  There is no longer complicated
memory managent of the command buffer.

What we lost: \c and \r commands.  \c can't be done in the current
setup because the data that you want cleared has already been sent to
the server.  \r could be done, but using \l to switch to line mode
accomplishes much the same thing.