Fix for a bug reported by Roberto todae (crash in a paralell batloop that is in a while-loop)
what happened here is that on the last iteration, if there are less than N
items, the MIL tree is hacked by the parloop code to hide some children
(for those statements where there is no data).

This hack should be undone afterwards, so the expanded tree that contains a MIL
parblock, stays in sane state, as itis re-used if the batloop itself is in
some other loop.