Fixed two incorrect uses of sizeof:
char *str = "NULL";
sizeof(str) -> equivalent to sizeof(char *), not to strlen(str).

sizeof("NULL") -> gives the size of the string *including trailing
NULL byte*.

Changed the former to char str[] = "NULL" with sizeof(str) - 1.
Changed the latter to sizeof("NULL") - 1.

This fixes a hang in MapiClient when a NULL result was returned
(MapiClient saw a nul byte and never saw the newline that came after.