Failure to allocate a large memory chunk in the context of BAT creation is not fatal anymore. An error message is generated and a 'zero' BAT is returned to its caller, which in turn may issue more errors.
The code changes propagete to GDKmallocmax and GDKvmalloc,
which now take an 'emergency' parameter to indicate whether
a memory allocation is fatal, or can be dealt with locally.

In GDKutils also an error counter (on private thread basis)
is maintained to simplify detection of any error occurrence
during multiple BAT creations.