(repeat msg as sf.net seems to have ignored it)
we allow persistent string heaps of BAT_APPEND bats to be unsafely mmapped (STORE_MMAP), after they stop to fdo perfect double elminiation.

(this is desirable, as such heaps tend to be very large, and saving a private heap would double disk space and consume much
 time due to unneccesary I/O on repeatinmg the stable part of the heap when it would be saved as STORE_PRIV)

However(1), this means that in a crash, the latter part of a str heap may be in inconsistent state. As all that
part is after the insert line of the last commit that didn't make it, this does not matter.

However(2), inserts also modify the hash table at the start of the heap. Consequently, it may contain pointers to
now garbage after the failed insert line after such a crash.

this seems an impasse. how can we proceed sanely or clean up this hash table? do we have to?

However(3), this hash table needs not be consistent so we don't care. We just make sure we ignore out-of-bounds
values encountered in it by considereing them end-of-hash-list values.