Fixed a number of related problems in Mapi.
If a multi-query command is incomplete (i.e. it needs more data to
make a syntactically correct query) but the first query is complete
and produces output, there was no way (due to the internal structure
of the library) in which the library could tell the application that
the server needs more data.

If a subsequent query in the command produces an error, there was no
way for the library to notify the application.

When the first query of an incomplete multi-query command was
finished, the library would send the "Xclose" command to the server,
even though the server wants more data to complete the query.

There is a potential deadlock in the interaction between server and
client.  This deadlock is not solved but circumvented in MapiClient.
If the client sends a huge command consisting of multiple queries, the
server will start executing the first query as soon as it is read.  If
this query produces lots of output, the server may hang sending the
data to the client, since the client is still sending data to the
server.  But since the server is not reading that data, the client
hangs as well.  This problem was circumvented in MapiClient by sending
queries in smaller chunks and reading output from the server between
chunks (but for this the first bug above had to be fixed first).

The function mapi_next_result can now return MMORE to indicate that
the last query was incomplete.
There is a function char * mapi_result_error(MapiHdl hdl) which should
be called after a successful mapi_next_result which returns any errors
from the server for the new result set.