Made internal functions static. Rearranged code so that no forward declarations are needed. Removed empty statement blocks. Removed includes from .h file. Added license/copyright. Made global variables static.
I don't really like the use of the global variables, but in the
context where this file is used (MapiClient) is doesn't matter so
much.  But if this file is ever used in a context where multiple
connections to multiple servers is possible, it breaks down.

I don't like the fixed file names for the history file.  A better API
would be to pass the filename to both the init and deinit functions.
The filename could be the argument of the -H flag in MapiClient or the
value of an environment variable.