All communication to the server is now done via the use of a (message)queue which is consumed by a single thread.  This results in the situation where only that single thread needs and uses the socket to monetdb.  No locks are needed on the socket anymore, since only one thread can and will access it in principle.  This is good on the once side as an expensive lock acquirement operation does not have to be applied anymore on a core level.  This is also good on the other side, where more scary things can be done, such as giving another thread access to the socket.
This change fixes the bug where the JDBC (and other tools as well)
locked up when sending large amounts of data at once to the server.
Thanks to Sjoerd we found out this has to do with the limited size of a
TCP stack.  For short, a deadlock situation occurs between the client
and the server, that both want to write to the stream, but block because
the buffers are full.

Now, JDBC uses a separate thread in case of large queries to send them
while the main queue runner thread checks if there is something to read.
This way we can send enormous strings, and read results as soon as they
come available.