Fix compilation issues with icc with respect to our use of inline.
The issue was that icc doesn't like it if you use static
variables/functions in an inline function with external linkage.

The fix involves making a bunch of static variables/functions extern
(all of these have names starting with the GDK prefix), and (worse)
using #define for a bunch of functions that were declared as static
inline in gdk.h.  The static inline functions have the advantage over
the #define'ed versions in that they are more readable, and they don't
require multiple evaluations of the arguments (as the functions can
use local variables).

The original static inline functions are still in the source for
reference, and will be used if the token
ALLOW_INLINE_FUNCS_IN_HEADER_FILES is defined in the C preprocessor.

Perhaps this should not be propageted to the stable branch.