Implemented the possibility to run the Mapi protocol over UNIX domain sockets (on systems that support this, obviously). There is a new, fundamental function defined in mapi.mil, called mapi_listen2 (*) which is like the old mapi_listen (which is still available, but now as PROC), except it has one extra argument which denotes the name of a UNIX domain socket file.  The listener started will listen to both the TCP/IP socket denoted by the port (if > 0) and the UNIX domain socket (if not "" or str(nil)).
On the client side, if mapi_connect's host argument starts with a '/',
it denotes the name of the socket file to be used and the port will
then be ignored.  (This works, since UNIX domain sockets are not
supported on Windows.)

(*) What I wanted to do is have a PROC mapi_listen with the calling
sequence of the old function, and a new builtin mapi_listen with one
extra argument.  But this is not possible, since it is not possible to
have builtins and procs with the same name, even though the arguments
could disambiguate.