my first direct pf check-in?
- close temporary files, please
- enclose all MIL in a block so variables are cleaned up
  (each xquery command leaked 60 bats.. before this)

I wanted to use pipe() iso files, but that would not work on windoze.

Better option is to convert to milprint_summer.c to Niels's streams.
Then we can collect the stream as a string buffer, and give it to monet_eval.

but well, for your average xquery with a MIL script of 100KB / 6000 lines,
the file overhead probably is not so big..