adds support for: (1) MONET_OID32, a define that enforces var_t and oid to be a unsigned 32-bits     integer, even on 64-bits systems. It allows to trade scalability for reduced data size.
(2) MT_mmap_trim, a new thread that flushes the (Linux) virtual memory maps
    when the RSS exceeds the MonetDB memory thresholds.


ad (1)

the changes concern mainly the correct use of var_t, where previously often size_t was used.
On 64-bits systems with MONET_OID_32 set, these are no longer the same.
MONET_OID32 is currently not set; we need a configure option to do that.


ad (2)

a list is kept with all sizable mmapped regions. It is used to keep the filedescriptors
to the files that are mmaped open. You look up a map by base address. The file descriptor
is needed to be able to call posix_fadvise(). This is done regularly by a background
thread, that is created automatically. This allows to do trimming even if MonetDB
spends all its time inside one command.

Background.. Linux has been observed to come to start swapping heavily as soon as MonetDB
uses a lot of (memory mapped) memory, even if the access patterns are sequential.
When in page need, Linux also drops all sequential readaheads; which severly hurts bandwidth.
The system (OS) even becomes unstable and crashes in many occasions (this indicates Linux bugs).

The cause of the swapping is that the file system keeps caching all pages of mmaped files
that were once loaded. On Linux 2.6.X, one must call posix_madvise(POSIX_MADV_DONTNEED)
*and* posix_fadvise(POSIX_FADVISE_DONTNEED) in order to make the filesystem buffer manager
drop these pages. Actually, one also needs to configure a low value of /proc/sys/vm/swappiness
(i.e. 0). This means that no unused memory is added to the filesystem cache automatically and will
hurt file server performance, for files that are reused. Regrettably this appears to be the only
way to make Linux stop caching files for which we specify negative usage advise.

With the current fix (and swappiness=0), MonetDB can keep functioning even when using more
memory than RAM. Operator speeds of course do drop to the supported sequential disk write
bandwidth (often a factor of 10 of performance slowdown).