removing code that worked around a non-existing bug in MonetDB:
The index of the first BUN in a BAT is not fixed to be 0 or 1,
it can vary, depending on whether we do have a hole for deleted
BUN or not. This is actually not a bug, but a feature ---
though a "legacy" one, since the hole is not really (not at all?)
used any more...

Anyway, "BUNindex(b,BUNfirst(b))" kindly gives us the index of the
first BUN, hence, we used this instead of a hard-code "1".

(Sorry for the "2 hrs", Jan...)