msec -> usec in sql_gencode fixed handling of sql prepare/execute
For example

prepare select * from tables where name = ?;

results in
1
1
varchar(1024,0)

Line one is the statement identifier.
Line two the number of arguments
Line three is the type of the first arg in the format type(digits,scale), where
we misuse digits and scale as we do in the sql types table.

So to execute a prepared statement

execute 1('ptables');

here the 1 is the statement identifier. Execute requires correctly typed
arguments.