Weekend Hax ;)
Adding some XML:DB files.  I stopped with the implementation because
it's a big mess in the interface specification.  This is non-working
stuff, just checking in to somehow preserve the work done.  I think the
best way to go now is create a piece of example code which just uses
JDBC directly, instead of a thin wrapper.  We can add wrappers on
demand, if someone knows a good one, currently there does not seem to
exist good ones.