fixes for the bugs
[ 1238867 ] SQL: PREPARE ... where 1 = ?
and
[ 1234205 ] SQL: PREPARE doesn't like LIKE

For simple prepares the type checking was done to late. So the parameters
type was never known, leading to a null pointer crash.

The second bug reported in 1234205 was about queries with literals.
In case of prepared statements these should not be put in the argument list
(that is only needed for statement caching).