fixing BUG [ 1267155 ] Pump properties:
The pump implementation (interpret_setaggr) correctly set
the key property for the head of the result.
However, if the result was the same size as the input
(i.e., the input's head was actually key/unique),
the code is "clever" and calls ALIGNsetH to propagate
the properties of the input's head to the result's head;
in case the input's head was not marked key (though it was),
ALIGNsetH overwrote the previously correctly set key property
of the result's head.

I fixed this, be setting the key property of the result's head
only after the ALIGNsetH call.