Removed autogen, Mx, mel, and burg from the MonetDB and pathfinder repositories.  They are now available in the buildtools repository.
To compile MonetDB and other projects, first check out, compile and
install the buildtools (available from SourceForge in a tree parallel
to MonetDB, pathfinder, sql, and xml), and then bootstrap, configure,
and make.

In order for bootstrap to be successful, autogen.py must be found in
your $PATH.
In order for configure to be successful, you must *either* have Mx and
mel (and burg for pathfinder) in your $PATH, *or* you can use the
configure options --with-mx, --with-mel, and --with-burg (they take
the name of the binary as argument, i.e. not a directory).

Also updated the HowToStart file.  The file is now written using
reStructured Text from which an HTML version can easily be generated.
This is done using a command rst2html.py which is part of a package
called docutils (http://docutils.sf.net/).