solves bug 1337630, which claims monet can't even join.
Problem is that merged_union produces an illegal sorted property
on the tail column of its first result bat (see example
script at sourceforge), in the second case, where the merged_union
is done on a [void,void] bat that is passed twice.

The problem is that when two threads, or (as in this case)
the same thread, intermittendly use the

BUNhead(b,p)/BUNtail(b,p)/BUNhvar(b,p)/BUNtvar(b,p)

macros on the same bat, but at different BUN positions 'p'
the last invocation overwrites the result delivered by the
former invocations (repeated invocations overwrite the same
field in the bat header. This kludge was adopted in order
to produce a pointer to an oid, which is difficult for a void
column that is never materialized).

Another kludge in monet_interpreter.mx already compensates for
the multi-threading risk posed by this. The safety mechanism
consists of making a BAT-view temporary copy if we detect
concurrent use of the same void-BAT by multiple threads. As a
BAT-view gets a private BAT header, this removes the BUN macro
problems (that stem from re-using a field in the bat header
to materialize the oid in).

Until now, the monet code developer was in charge of ensuring
that this would not go wrong in the single-threaded case.
Thanks to full materialization, this boils down to the case
were the same void-BAT is passed multiple times to a command.

i.e. todays bug.

Thus, all MIL commands must make sure that those BUN macros
are never used on *different* BAT parameters *without* copying
the accessed value to some temporary variable.

Alas, if even Stefan can't handle this, we must conclude that
leaving this to the developer is too optimistic.

The fix of today extends the multi-threaded safety mechanism
(interpret_pin) to the single-threaded case of a multiply
passed bat.

open question: Monet5 developers are you even aware of this issue?