- constant marshalling is now fully transparent thanks to overloaded   insert(bat[any,bat] b, any h, constant t) => insert(b,h,constant2bat(t))   fetch(bat[any,bat] b, int pos) => fetch(b,pos).bat2constant()
The only explicit use of constant2bat() is when invoking print_result()
and the loop-lifted Xpath steps (which require BAT params).

Thanks to this change, we can be pretty sure at all times that
the (iter-pos-item-kind) bats either contain a full BAT[void,T]
or a MIL constant. No more fake-projects to worry about.