BugDay_2005-12-19, Jennie: TEST ADDED / SUCCESS
src/sql/src/test/BugDay_2005-12-19_2.9.3/Tests/prepare_doesnot_like_LIKE.SF-1234205.sql

Note: the test script is added to show "prepare select * from env where
name like ?" works (correctly).  The ID of the prepared statement is
hard coded in the test script, so the test can fail if the prepared
statement gets another ID.

A test script in application level should be added to dynamically
retrieve the ID of the prepared statement and call 'execute' with the
right ID.  But this is out of the scoope of the BugDay.