Supply client with more data (that it eventually needs). the table_name header now consists of "schema.table" fields, instead of only "table".  This allows the client to extract the schema name.  I chose not to add another header because from experience we know that is bad for performance.
Only problematic thing on this commit is that I don't know how to get
the schema name out of a statement in all cases, and I fear that this is
impossible for many cases.  In the meanwhile I hardcoded schema "sys"
for some cases in order to allow the client to play with it.

Note that a table_name field now should *always* contain a dot!