Inclusion of experimental code to recycle BATs. The code is not activated though, pending peer review.
In bbp the physical location of the BAT is recycled,
because this information only depends on the BBP entry.
This avoids a lot of malloc/frees, making createdesc about
twice as fast and on the DBTest it gains another 3% total.
A similar step can be taken with BBP_logical, but requires
a more careful code check. It is also (mis)used to determine availability
of the entry. [another 2% gain would result]