The check in GDKfree detects obvious duplicate free's, but fails in case the "check-bit" is cleared between two free's (e.g., as the respective memory has been re-allocated and initialized.
To simplify detection & debugging of duplicate free's, we now overwrite the
to be freed memory with "invalid" data, which will (1) trigger a segfault in
case the memory had already been freed and/or trigger some error in case the
memory is accessed after is has been freed.

To avoid performance penalty in the "production version", we only do this in
debugging/development mode.