[ 1434079 ] CTrefine failes in xtables tests
the last 16 bytes of a heap that just borders on a 64KB boundary where lost.

this bug has been in place for a long time..

it was exposed by setting the VM allocation threshold lower, and the xtables
scrips allocating bats of nice powers of two (which since the disappearance of
the BUN hole translate in heaps of nice powers of two).

let's say thank you to regression testing and those who support it (StM)