Loop-lifted inserts.  In any case, partially.
We still do a loop per insert command, but we no longer do a loop per
to-be-inserted node (and within that loop a loop per to-be-inserted
attribute).
The code to make space for inserts is unchanged, the code to copy the
data into the newly created space was completely rewritten.
We now do a loop per page, and on each page we do bulk operations to
do the work.
Also removed some procs that are not needed anymore since we now have
bulk versions that we use.

Timing of a test insert into the beginning of a two-and-a-half page
document:
before_swizzle: 5.3 seconds
after swizzle, before loop-lefted insert: 8 seconds
this version: 1.8 seconds.
Note that this is a pretty unscientific test.  ;-)