Added sql_trace (m4)/sql.trace (m5) functions, these enable/disable trace logs of sql clients. So a client started after an odd number of sql_trace calls will log all statements to a file in the dbfarm (called sql.'client_id').
Added (to 4 only for now) mvc_append, this is the combination of a mvc_bind
and append call. This way we can replace the insert bat by the to be inserted
one if this is possible (ie head type void and old inserts bat empty).

approved some tests after these and other recent changes (mostly \" arround
table/column names and trigger table related)