Removed the compile-flag "--enable-burk", and made it a runtime option "--enable-standoff".
last-minute... i know (it is a little past the deadline, but hopefully still accepted) I have put tags around the checkin so that the modifications can be reversed.

This change has a subtle influence on error-reporting by the XQuery-compiler. The change is that when --enable-standoff is NOT used, an XQuery trying to use the StandOff axis steps will not get the "invalid character" error, but "invalid character, (StandOff was not enabled)".

The StandOff axis steps still do not work with the new indexing scheme, but that's a bugfix, not a feature ;)

And the comments I left in the code are pretty ugly, I'll clean them up the upcoming days.