Two new types are introduced: bte and wrd.
bte is an 8-bit signed integer type.

wrd is a system-dependent signed integer type, 32 bits on a 32-bit
architecture and 64 bits on a 64-bit architecture.

TYPE_str is still the last built-in type, but it now has a different
value.

There are tags before_bte_wrd and after_bte_wrd around this checkin.

Stefan did much of the work, I finished it.