UPDATE YOUR buildtools!!! UPDATE YOUR MonetDB!!! THEN bootstrap/configure/make/make install IN MonetDB!!! THEN bootstrap/configure/make/make install OTHER PACKAGES!!!
We add an extra directory level for dynamically loaded modules (.so
files or .dll files).  This is to solve a problem on Cygwin where
there was a name clash between MonetDB 4 and 5.  The original solution
of using the -module flag of libtool caused messages that linking to
modules is not portable, and indeed, on Darwin that didn't work.

This change is made in autogen and so requires an update
(bootstrap/configure/make/make install) of buildtools, MonetDB and all
dependent packages.