Major signature change. The mal_io primitives were all marked as unsafe functions, i.e. they produce side-effects. This is conveniently derived from the function return type. A :void means that something changed in the environment or in the arguments itself.
This lead to making removing the 'unsafe' property and
returning a :void instead. Errors during io are normal
exceptions.

The consequence of this change is that about 297 tests
had to be re-approved. They follow.

At the same time, the use of GDK_FAIL and GDK_SUCCEED
is further reduced, relying on exception handling instead.