stmt_output is now added in output_semantic this makes it possible to call semantic also for subqueries which is used by scope_sqlparse
now the same functions handle creation and instantiation of views,
functions and triggers, ie the later with mode == m_execute.
So from now on we store the full sql code with views, functions and triggers.
(this change was needed to handle triggers with a list of statements in the
trigger action)

recursive called triggers are now protected by a list with in the mvc structure
(called_triggers).