This commit breaks 4 tests.  I do commit, however, because SQL needs it.
I ran into the problem that in mal_interpreter basically two exceptions
are molded into one.  This results in the problem that there are two
locations where the exception happened, and Martin just glued them
together.  Conceptually this is hard to grok, and since it's not
consistant, it's hard to parse.  It needs more thinking whether we
should (re)introduce a stack trace, or that we extend the location field
to have for instance "user.main[74]-bbp.bind", or something else, which
is parseable and consistent and useful...