our code has become cleaner and/or the "problematic" tools obsolete; hence, it's time to clean-up our (X_)CFLAGS extensions.
to detect portability problems as early as possible,
we try to make gcc as strict/picky as icc, MSVC++, etc.
(only when configured with --enable-strict; default in the development trunk);
in particular, disallow pointer-arithmetic on void pointers
and ensure that declarations come before statements

Note:
Please update your buildtools
and re-bootstrap the code that you are currently developing on.