Changes to support: 	[DELETE | UPDATE] [CASCADE | RESTRICT (default one) | SET NULL | SET DEFAULT | NO ACTION] 	DROP [CASCADE | RESTRICT (default one) ]
Some modifications were:
	to distinguish Foreign key from a unique\primary key
	to find a database object using the id
	to be able to call some function recursively (new argument, schema, was add to some functions)
	a new column (called action) was added to the table keys
	some of the sql_trans_drop_* functions received a new argument to be used in the recursive call
for the cascade drop


Note: The existent test were not broken by this new code. However, the implementation itself
needs to be tested which will be the next step. The commit could not be done earlier due some problems
in my machine. Maybe huge amount of changes will create conflicts in your code, I hope not.