fixing BUG #1706438 "PF: property error when adding a doc. to an existing coll."
Though we do not set the key-ness property in the shredder,
property checking (BATpropcheck) might have detected and set the
key property that "happens to hold" after an initial shred of a
single document. When added more document(s) to the same
collection, we need to make sure, that we unset the then (most
probably, if not surely) no longer valid property.
(Since we usually do not expect and hence exploit the key property
in the size & kind BATs, we omit an expensive check to find out
whether it might still hold...)