MapiClient has been extended with a SQL rendering option based on a predefined columnwidth and pagesize. A columnwidth= -1 means old display structure. Columnwidth=0 indicates to respect the sizes in the table headers. Columnwidth>0 is an upperbound. Broad columns are pushed to make all fit. The long values are displayed over multiple lines.
The pagesize default is -1 to indicate the old situation. A pagesize
= 0 indicates no pagination structure, and a pagesize>0 produces
a paginated result, which can be stopped.

The styling works with the debugger, which now also better indicates
the mode. The trace mode requires some extra work (for columnwidth>=0)

The style conventions for the other languages should not have been
changed. (checking is needed) (and multi platform compilation)