First step to have partition distribution. The distribution is done at the partition fragment level. Each fragment resulted from a bat partition can be located at a different location than the actual bat location.
For the moment the location information is (by default):
	host (localhost)
	port (50000)
	dbname (demo)
	user (monetdb)
	passwd (monetdb)

The first three attributes are used to distinguish different locations.

This file contains storage management for the location information.
	add location
	remove location
	find location

The load of all the information once the Mserver restarts is also implemented.

Note: It was compiled with strict compilation and no errors showed up. For the moment
no tests were ran against this code.