This is a bug fix for: [ 1881018 ] PF: value clash in compiler/include/algebra.h
My attempt to use less bits in the type system of pathfinder did not
work. Since, we also agreed that at some point we should extend the
type system, I end up using 3 more bits to make it work.

I guess we can close the bug or wait for the night testing and then close it.