Fix for bug #1913438 (Merovingian: no connections accepted untill first command)
Since I made the control channel non-blocking, the open() call blocks as
well.  This prevents merovingian from reaching the code where the TCP
sockets are handled until the first control command is sent, which
unblocks the open() call.

Instead, now block in the control runner thread itself.  This thread is
made for waiting on input ;)