A fix for bug  [ monetdb-Bugs-1961027 ] (and might be a more general fix or a more general mistake:))
In algebra_cse.c, in fynction subexp_eq the case la_lit_tbl and la_empty_tbl is treated the same.

However, in my understanding empty_tbl does not have any tuples, so if the schema.count and schema.items
are equal, then the case of the empty_tbl should return true, otherwise continue the comparison of
tuples since it is a lit_tbl.

Please comment on this if it is correct.