Fix for bug [ monetdb-Bugs-1964365 ] PF/Alg: insert..into inserts in wrong order
The order of a sequence parameter (that is inserting a sequence of elements in one insert statement and not
a sequence of multiple inserts) was not respected by the algebra until now. The reason for that was that the reversed
pos column was not projected back to the result relation.