Got rid of descriptor files.  This has a number of consequences which are detailed below.
First of all, this checkin requires you to do a full
dump/delete/restore if you want to keep your data.

Also, this checkin requires to do a full rebuild of all components
(except buildtools, clients, java).

Instead of writing the descriptor to disk, we now write the relevant
information in the BBP.dir file.

Since the descriptors are no longer saved on disk, they have to be
kept in core.  In order to reduce the amount of memory, the BATstore
structure was extended with an area to keep the BAT structure for
mirrored BATs.  This means that we can save on one malloc with malloc
overhead.  This does not change the way the mirror is accessed or the
way the BAT pointer and its mirror are stored in the BBP table.

Also, various bits that were stored in full bytes (or more) were
coalesced into bit fields.  In particular, the values set, restricted,
persistence in the BATrec and key, nonil, nil in the COLrec.

The format of the BBP.dir file has changed drastically to contain all
information that needs to be saved and would otherwise have been saved
in the descriptor files.

There is now also a version number in the BBP.dir file.

There is no longer a version number in the BAT descriptor.  There is
no need, since it wouldn't be saved on disk anyway.

Some internal functions got a signature change, mostly changing
"pointer to BAT" to "pointer to BATstore" (in either the arguments or
in the return value).  Affected functions are: BATnewstorage,
BATcreatedesc, BATdestroy, BBPinsert, BBPcacheit, BBPrecycle,
DESCload.

The following files were deleted: gdk_desc.mx and upgrade.mx (MonetDB4
module).

Only two small changes were needed outside of the core.  Both M4 and
M5 use BATcreatedesc and BBPcacheit in one place.  This use had to be
changed.  An incorrect usage of the TRANSIENT value was corrected in
MonetDB5 (it's a value, not a bit mask)--important since the TRANSIENT
bit pattern has changed since it was put into a bit field.

An interesting bug in subcommit was also found and fixed: in the old
situation when you have a committed, persistent bat that you rename,
then the changed name is committed in the next subcommit, even when
the bat itself was not subcommitted!

I ran tests in MonetDB4, MonetDB5, and sql.  The test results were no
worse than before.

Before tag: Before_descriptorless
After tag: After_descriptorless