Redo of fix, but this time with one fewer bugs:
This should fix bug 2141012 XQ: unclear error.
We now look at the collection size as stored at position 0 in the
PRE_SIZE table to determine where the next document should start
counting.  This value results in a PRE value which we translate to a
RID value.
We do not have to do anything with NIL values at the end of a page
which were put there to make sure the page was completely filled,
since we don't use the BATcount to determine where to start.