Changed password store from plain text to SHA-1 passwords.
Basically this whole change was necessary for SQL user dumping without
plain passwords.

As of now, we need OpenSSL for SHA-1 algorithm, mandated by Mapi
protocol v9 to hash the password.

Upon first load, the old password store is converted on-the-fly, to
guarantee a clean upgrade path from the previous release.

Implementing the SQL dump stuff is the next step.