fix for the lost update bug.
The update got lost because the logger_restart saved the current
bat (with the update) but also the udate bat (old version). Many
small problems, but most important is that BATins on a Bound to be key
bat doesn't replace the values for the last but keeps the current values.
We now simpl forget about the update bat and only keep a single persistent
bat, ie sql now also uses the log_delta, once created for xquery.