Improved Martin's previous fix.
With this fix, the test in bug 2791361: "Division by zero error" and
the Periscope test CASE_division_by_zero give results (the former
gives the correct results, for the latter the results need to be
checked).

However, this fix changes the semantics for DIV on bats somewhat: if
you now attempt to divide by zero, you get a NIL whereas previously
you would get an error (which is what caused the problem in the above
mentioned tests).  Thus, this fix needs to be checked for its
correctness.