Time to show what I'm hiding.  (Also nice to finally get revision control)
My efforts of the last few days resulted in this initial version of
"quicktest", the not-even-close-to-Mtest.py fast mtest emulator.

Basically it started all out of frustration of not being able to
implement many ideas in Mtest.py, so I decided to see how far I'd come
by rewriting from scratch (incidentially also in a language I actually
can read and understand).

All it can do up to now is just .mal tests, and it will fail on all of
them because it can only (and will only) use mclient to run them,
causing diffs in the output.  Correcting those actually make it able to
also report success.