The fragment limit in the new fragmented index is based on number of used PREs, rather than documents. So, it is not known a priori how many documents will fit in one fragment. For this reason, the fragmented indexing was done per-document. After each document the fragment size was checked. This introduced two important reasons for bad performance and scalability:
1) [performance] a working set was created for each document. This cost far more than the document indexing itself
2) [scalability] a dbat set was created for each document (using _tj_throw2collection instead of _tj_throw2collection_bat). This triggered a dbat_extend for each document inserted, with terrible consequences as soon as the bats were memory mapped.

Fixed by indexing in batches of documents, so that the overhead is amortised. For very large documents, this may not be optimal, because the fragment limit would be checked too rarely. By setting documentSize="large" in <TijahOptions>, the batch size is set to 1, which reproduces the behaviour of one document at the time.