Changed decref to first check (with the sapLock held) whether the bat to be decref'ed is unloading, and if so, to wait, instead of first doing the decref and then, when the bat was unloading, to do a temporary incref.  This (hopefully) resolves a race condition.
Changed BBPreclaim to just call decref, since that is basically what
it does.  Added an assert to make sure BBPreclaim is only called when
it is allowed, that is to say, when the bat was privately owned by the
caller (refs == 1, lrefs == 0).

Changed some calls to MT_sleep_ms to calls to BBPspin, since that is
what was being done.