make sure we do not leak:
ATOMfromstr() creates a buffer (or reuses and if necessary extends the one provided)
to store its result;
ATOMput() copies the result from that buffer into the BAT's BUN- or string heap.

Hence, once done, the temporary buffer need to be freed in all cases,
and not only for "extern" (varsized) ATOMs.

(In other places with similar code, we need to free to buffer
only for NON "extern" (fixsized) ATOMs, as there the result
of ATOMfromstr() is placed into a value record. Only for
"extern" (fixsized) ATOMs, the actual value is copied
(and hence the temp. buffer needs to be freed afterwards),
while for "extern" (varsized) ATOMs the pointer to the buffer
is assigned to the value record, and consequently, the buffer
needs to be kept for now, to be freed once the value record
is freed.)