Perform typecheck before registering the function at the remote side.
It can happen that a function has not yet been typechecked when it's
sent to the remote side.  Since the materialisation into a string then
loses the vital typing information (that is deduced from the plan), this
results in an invalid plan on the remote side.  Hence, check the plan
before sending it, such that materialisation contains all information.
If the typecheck yields errors, we throw an exception (this is a win: we
catch errors early this way).