Express dependencies between bat.new and bat.insert operations for dataflow.
Make a chain of references from bat.new to all bat.insert calls, such
that the order between them is explicit.  This should help dataflow to
figure out correctly what can be executed independently and what not.

In this code this change wasn't trivial, since in case of a "count"
optimisation we need to traverse the chain and fix types here and there.