Try and solve autoloading problem as used by testing.
Refine the recursive searching for autoload modules not to stop on the
first found directory, but to keep on scanning in the mod_path for as
long as there are entries.  In this way all modules found in the entire
mod_path are returned, instead of only the ones in the first directory
found in the mod_path.

Ordering is based first on mod_path order, then in each directory sorted
by name.