Tool to compare the results of two different Mtest runs.
This tool compares the "times.lst" files in the two given
directories and reports all tests whose results differ.  If the
standard output differs, differences in the error output are not
reported.  The tool is meant to find tests whose results differ, not
to list all differences.

Optionally, significant (more than 10%) slow down of succeeding
tests can also be reported.  (If tests fail, the speed is useless
and thus not reported.)  Note that comparing times is only useful if
the tests were run under comparable circumstances (optimization,
machine load, etc).