refining changeset 647c99679171 to avoid performance degradation:
changeset 36465:647c99679171 said:
"
avoid hash overhead when looking up group IDs in group extend

Using hasing to lookup group IDs in the group extend --- basically
performing a foreign-key join between group map an group extend ---
is rather expensive, not only because both creating the hash table
and probing the hash table inherently require random access, but also
because the "hash table" (mask & linked list) are rather large.
However, in practice, the group extend is often sorted, and the group
map "almost sorted" or at least clustered such that group IDs in
cluster i are smaller than those in cluster i+1 (this is the case if
the first grouping attribute (column/BAT) is sorted).
In this case, performing a "layman's merge" by keeping a cursor
into the group extend that simply performs a local forward/backward
scan starting from the position of the last lookup, turned out to
perform significantly better than a hash lookup.
"

However, the assumption that with the group extend BAT (e) being sorted
on head also the group map BAT (g) is "almost sorted" or "mostly ascending"
or at least clustered such that group IDs in cluster i are smaller than
those in cluster i+1, does not appear to be true in all cases.
In case the group extend BAT (e) is sorted on head but the group map BAT (g)
is not "almost sorted" / "mostly ascending" but rather randomly ordered,
the newly introduced "layman's merge" performs sugnificanly worse than
standard hash-lookup.

Lacking a suitable and efficient way to detect that the tail of the group
map (g) is "almost sorted" or "mostly ascending", for the time being, we
deploy the "layman's merge" only in case the tail of the group map (g)
is fully sorted (which is most probably only the case in case the table
is sorted on all group-by columns) to avoid a performance degradation.

The "layman's merge" optimization will be re-enabled once we have a
suitable and efficient way to detect that the tail of the group map (g)
is "almost sorted" or "mostly ascending".