Work around UNIX domain socket length restriction of around 108 chars.
The sun_path element is typically around only 108 chars long.  This
causes paths to be truncated.  Find below list for the actual limits on
sun_path and what it accepts in reality.  I worked around the problem by
using a relative socket file location (in the current working
directory), such that the long absolute path is no longer necessary.
For merovingian this wasn't much of a problem because it already chdirs
into the dbfarm, but for monetdb it required a chdir to dbfarm first.
It only does now if the connection to use is a UNIX domain socket.
Since all actions are done remotely, the working directory for monetdb
doesn't matter at all.

Solaris:
    defined sun_path = 108
    max sun_path = 1024

FreeBSD 8.0:
    defined sun_path = 104
    max sun_path = 254

Fedora 11:
    defined sun_path = 108
    max sun_path = 108

AIX 6.1:
    defined sun_path = 1023
    max sun_path = 1023