Fix for Bug #2641
When we encounter an Unicode BOM, skip it.  mclient does skip the BOM
when it happens to be the first thing in a file, but not all clients are
mclient afterall.  Hence, when e.g. a Java API application sends the BOM
it got an error.  The implementation in the SQL server is more complete
than what mclient does, though, as it also skips BOMs that appear in the
middle of a file.

I won't remove the BOM skipping from mclient, since it seems to be added
there for MonetDB/XQuery, which for sure doesn't do this BOM skipping.