With the SQL formatter, we used to silently ignore any random garbage the server sent us.  Change that from now on.
When random crap comes to us (Mapi unfortunately doesn't detect and
scream loudly about it), make sure we don't try to render it as a table,
but instead just tell we don't know how to render the garbage.  Note
that this only applies to the SQL table rendering mode.  As a result, we
fix the timer result on result sets not to output negative times due to
the client having absolutely no idea what it's displaying, and hence
having no means to stop the timer at the right moment.