Add gdk_testing environment variable.
GDK_testing is used to switch code on/off that is meant to be used
during testing only (an example is splitting small tables in the
mitosis optimizer to test mitosis/mergetable).  The value is set on
the command line using --set gdk_testing=true by Mtest.py.

With this change, small tables will not be split by mitosis (unless
gdk_testing is true) so that outside of testing, we don't get the
(sometimes enormous) overhead of split tables (the example that
triggered this went from 9.2 seconds to 46.827ms, almost a factor
200!).