Build a self-containing libmonetdb5
Rework building process to have a libmonetdb5 shared library which
includes all functionality that is loaded/necessary by a default
mserver5.  This way, embedded clients can link against libmonetdb5 and
need not to link against the private libraries of MonetDB5.

This also reduces the need to dynamically load many loadable objects,
since they are in the used libmonetdb5.

In this process, all pieces of code that need to be loaded dynamically
(e.g. are not included because they are optional), were moved to extras.
This is necessary, because they can only be built after libmonetdb5 is
built, since they have to link against this library.