change installation of files
Introduced CONDINST for binaries, such that tests need not to be
installed when --disable-testing is given, but make check still works.

Only install headers that are actually useful for an external consumer.
In principle, these are only the headers that are connected to a library
that we install in ${libdir}.

Only install public consumable libs in ${libdir}, the rest under
${libdir}/monetdb*/(lib)?.

Install public headers in include/monetdb.

Lowercase the directories being used.