multiplex: properly null-terminate stream read buffer
In particular JDBC doesn't send newlines in requests, causing a buffer
to just end without, and no end of string marker.  This yields in the
severs seeing lots of old garbage, on which they complain.  Switch from
mnstr_read_block to mnstr_read, because the former doesn't properly set
the read length for some reason, which we need to properly terminate the
block read.  Ignore implicit flushes (reads of size 0) that the protocol
inserts on final blocks.

This allows to query using different queries (different sizes most
importantly), which helps clients that do a lot of meta-queries to
initialise and retrieve additional data.