merovingian: implement dbfarm oriented mode of operation
Too many changes to list precisely, a combination of cleanup, refinement
and making merovingian work with its own properties that can be set
using merovingian itself.  This is all but completed yet, but this is
already an improvement over the current situation which is more broken
than the state of affairs of this patch.

This patch has some repercussions on other pieces of code, in particular
mapi and the monetdb client tool.

- extend confkeyval with an int version of the value for easy use with
  numbers, add convenience wrappers for retrieving values for a key
- properly disable discoveryservice, as advertised in the manpage
- .merovingian_properties files are only used by merovingian these days,
  a template that ensures backwards compatability for systemwide
  installs is installed
- removed replacePrefix function as it is no longer necessary
- unix sockets are now placed in /tmp by default, and named
  .s.monetdb.<tcpport> for mapi connections and .s.merovingian.<tcpport>
  for control connections
- monetdbd now understands actions (like monetdb) 'create', 'start',
  'stop', 'get' and 'set' that operate on a dbfarm directory
- logfile capabilities have been reduced, now only a single logfile can
  be used, and logging to stdout/stderr is no longer possible