loadLibrary: simplify dlopening of libraries
Rely more on dlopen to load a library, instead of first checking if a
file exists, before opening.  Unfortunately, we cannot just give dlopen
the base name of the library we want to open, such that it can figure
out the file extension itself, because we don't have a search path setup
for our module dirs.

On Darwin, a module (bundle) is something distinct from a dynamic
library, and hence also has a different extension.  This patch makes
finding modules on Darwin working again, which means the server can find
its modules again and startup.