configure: remind user that we require GNU Make, in case make does not appear to be GNU Make.
Added simple check at end of configure, whether make is GNU Make.
If so, fine.
If not, but gmake is GNU Make, remind user to use gmake.
If neither make nor gmake appear to be GNU Make (or not available at all),
remind user that we require GNU Make to build MonetDB.