memory: drop getrlimit code block
The getrlimit approach was meant to detect memory-limited environments,
however, in practice it never does what it was supposed to do (except on
Solaris).

On Darwin, the data size is by default set to 6MiB, causing a default of
6MiB memory to be detected.  We can in reality allocate much more than
that.  The OS uses it as hint for victim choosing when memory shortage
occurs anyway.  On Linux the limit is never respected.  In short, nice
idea, but it doesn't quite work out as intended, and only seems to
confuse.