mnc: "M" n(et)c(at) -- read and write from/to sockets
This limited nc variant is a side effect of a benchmarking of our stream
library performance.  The outcome of the benchmark is that we are as
fast as nc and nfs are in moving around raw data, so there is nothing to
worry for us.  Since the tool in itself is quite ok and handy when
trying to talk block-mode to servers (something nc and telnet can't do),
I'll check it in, but as NOINST, as I don't think it should end up in
any release as installed product.