remote.get: fix subsequent binary get calls
A side effect of making remote08 test more testing of different pointers
revealed that there was left-over garbage on the stream causing
subsequent get calls to claim incorrect headers were sent.  Make the
output better in this case, and fix the double flush that was the actual
cause of the problem.