JDBC: add so_timeout Driver property
On case the server locks up, the JDBC remains "hanging" too, waiting for
the server to send something.  When the so_timeout property is set, a
timeout for this can be set, such that the read eventually breaks, and
the caller (application) can continue.

Based on original patch by Philippe Hanrigou in bug #2828