I added an sql operator called sample. It goes through all the sql stack with a new relation operator called op_sample (and the equiv. st_sample etc. ofcourse). Currently there is no m5 module sample commited yet so if you call SAMPLE you will get an MAL exception. However explain and plan should work fine. Some known issues right now are some mismatches on the optimising level, (failing to apply dce in some cases) but they will be fixed soon.
The following text is a more formal explanation which can also be used for the
"new features" text. Another commit in m5 to come later will discuss the
algorithmic aspects.

-------------------------------------------------------------------------------
A new SQL operator has been added to support sampling the result of a query.
The syntax for sampling is:

SELECT ... FROM ... WHERE ... SAMPLE sqlINT

where sqlINT is the number of rows in the sample. SAMPLE is been treated the
same as LIMIT, ORDER BY, etc., that means that it can only be in the outer most
SELECT clause, i.e., SAMPLE cannot appear in a subquery. However, if this is
needed, then one may define a function, for example

create function mysample ()
returns table(col a,...)
begin
	return
		select a,...
		from name_table
		limit sqlINT;
end;

and then use function mysample() for example to populate a new table with the
sample. E.g.,

INSERT INTO sample_table (select * FROM mysample());
------------------------------------------------------------------------------