Backed out changeset 2d9beef3e6f5
This change was done at the wrong time on the wrong branch.  In a very
late state it changes external interfaces.  It causes dumps from older
versions of the database to be interpreted differently (unacceptable)
and it causes incorrect results in ODBC (and likely JDBC) when dealing
with the changed types (unacceptable).
The fact that dump output had to be changed (added zeros to interval
types) should have been a dead giveaway.
I have propagated the change to the default branch where the kinks can
be worked out.