m5: turned MAL_MAXCLIENTS into a runtime variable
While 64 may be enough for most cases these days with single clients, it
certainly isn't for environments where hundreds of clients are "shot" at
the database as seen with e.g. the LOFAR pipeline.  Given that dataflow
also eats threads per core, on modern hardware the 64 client slots are
easily halfed.
Hence, allow the user to start the database with a max_clients setting,
to easily override the default 64 in cases where this makes sense.