Mtest: don't treat "Connection terminated" as segfault
The client cannot possibly see the difference between a server closing
the connection because it crashed or just because it was shot down (e.g.
by Mtimeout), so don't treat a disconnect from the server as a server
crash.

This resolves many crashes in the testweb, that in fact turn out to be
just timeouts.