sabaoth: fix running status for "own" databases
Since MT_lockf always succeeds when the same process already holds the
lock, a database cannot be determined to be crashed based on the result
of MT_lockf, when the database is looking at itself.  This scenario was
handled by simply assuming that the database was running when it
happened to have the same name as the one that sabaoth was inited with.
For databases this is in general fine.  However, when pseudo objects,
are registered like databases, and/or a single process has multiple
databases, this approach breaks down.
Each process that initialises sabaoth now, will generate an (unique)
UUID, that sabaoth writes in each database directory when it updates the
uplog file (registerStart).  This way, getStatus can check per database
it sees, if the process associated with that database happens to be the
same as the one currently performing the getStatus query.  As such, it
can determine the current running state based on the uplog file, with
the assumption that the process is still alive (or else it wouldn't be
querying at this point, would it?).