SQLparser: make error reporting to the client work (again)
Throwing an exception here isn't going to end up in a message to the
client, in the worst case it even makes everything hang.  Make sure we
properly inform the client about the problem we encountered.  This
solves in particular the problem that EXEC 3627(); on a fresh connection
returns without errors.