Implemented sorted and revsorted properties, but rid of radix sort properties. This is very much still work-in-progress (hence the branch). The old sorted property which encoded either GDK_SORTED or GDK_SORTED_REV with some bits left over for radix has been replaced by two one-bit properties: sorted and revsorted. Also new is a function BATassertProps(BAT *b) which checks properties and uses assert() in case a property is set incorrectly (i.e. set when it shouldn't be, *not* unset when it could be).  This function is called from BBPkeepref to assure that properties are set correctly when a BAT is returned to the MAL interpreter.  The function will *not* set property bits that it found could be set.  It also doesn't fully check the key property (it'll do a single scan, so keyness is only checked for sorted BATs).
In going over the properties, and especially after the introduction of
the BATassertProps function, it was found that in many places
properties were not set correctly.  These were fixed when encountered.

Currently the monetdb5 tests work pretty well (some differences, but
mostly benign).  The sql tests don't run correctly yet (more
properties that are set incorrectly).

By the way, I removed radix.mx since it has been unused for quite some
time.  If you need it, you can find it in the history.