snodgrass: enable all tests
Provide stable out/err outputs, and disable tests that are bound to
fail.  We might want to eventually check whether we need to add
parsing/execution support for the missing constructs and fix the invalid
outputs.
However since this is very low priority, it is much more useful to test
what works, and not have errors, so we don't have errors we won't fix.