Reinstate GDKvmalloc.  This is a partial backout of changesets 5ee4cc736169 and b6e70864037e.
GDKvmalloc and friends are now static functions in gdk_utils.c.  They
are used on non-glibc systems for larger allocations (default: >= 1MiB).

This is hopefully the final fix for bug 3067.

On Solaris we found that memory fragmentation became a real issue.
This is solved partially by using libumem (previous changeset), and
partially by using anonymous mmap.  Memory fragmentation becomes an
issue when a large chunk of memory is to be allocated, but the system
break (sbrk()) cannot be increased anymore, e.g. because too much
memory has already been committed to processes by the OS, and no large
enough contiguous piece of memory is available on the process heap.
Memory mapped files (anonymous or otherwise) don't use the same
address space and so don't increase the break.  The only concern then
is that enough memory is available in the system, but since all large
chunks are memory mapped, we don't waste as much space on the process
heap.  Also, when the memory is freed, the memory is returned to the
system and not kept allocated to the process as part of the process
heap.

On Linux (and presumable any other system that uses glibc), malloc
itself already uses this same trick for larger allocations, so we
don't have to do it.