mtime.c: made code work, again, after recent BAT-split (changeset d3c1ae10f807)
Listening to compilers and (nightly) testing:

Static variable "timezone_def" was declared and set but never used,
which either means that (1) it is correctly not used and thus its
declaration can (and should/must) safely be removed completely,
or that (2) its intended use has not (yet) been added to the code,
and thus the code is incorrect.

In this case, 20+ failing timezone-related tests indicate that
the BAT-split that introduced static variable "timezone_def" (and
"timezone_name"; see changeset 92da8714361d) was not working
correctly, yet, i.e., (2) is likely to be the case.

And indeed, listening to the hints given by compilers and (nightly)
testing, a careful look a the code reveals that "timezone_def"
rather than "timezone_name" has to be used in function "MTIMEtimezone()".
Thus, the "fake use" of variable "timezone_def" is indeed not
the correct solution.

This fixes the 20+ timezone-related tests that have been failing since
changeset d3c1ae10f807 "Split the timezones catalog".