unwrap: fix handing of empty arrays
Don't just overwrite previous results when an empty array is
encountered.  Multiple arrays are possibly in the array being checked,
hence there can also be empty ones in there after non-empty ones.  This
also reveals a thinko: the type of the empty result doesn't have to be
lng, because it doesn't have to be the only array processed!