casts: slightly change to trigger less CPU-dependent differences
Avoid being on the edge of what an int or lng can hold, but go far
beyond it, such that representations on different CPUs (most prominently
big-endian CPUs) don't trigger the overflow that was anticipated for.

Hopefully this makes the test run fine also on PowerPC (vomac/Darwin,
macbeth/Linux) and should remove the need for SunOS output (Sparc for
real).