BBP: increase size (allow for more BATs)
BBP is a fixed-size array of pointers to arrays of BBPrec structures.
The size we have chosen for it, seems too small for some extreme cases.
We increase the size of the array of pointers by factor 10 (to 1000),
and on 64-bit systems increase the size of the array with BBPrec
structures by factor 2 (to ~16K).