jaql: undo operation reordering
We used to "optimise" operations such that variables would always be on
the right, making it easier to work with.  However, although this works
fine for addition and multiplication, it changes the meaning of negation
and division.  So, better not reorder arguments for this, and simply
support all combinations.  This turned out to be trivial to support in
the code generation, but required a fix to VIEWreset.
Enabled the calculations test that revealed this problem.