dumpvariabletransformation: fix using a possibly unrelated var
The function calling code used the (last) variable used for calling the
function as translation between returned ids from the function and ids
for the elements from the input array.  This would break when the
function had more variable arguments, or when there would be no variable
argument at all.  In the latter case a crash would be the result.