Introduce a sink point for parallel processing When a variable is used multiple times in a dataflow block, we have to ensure that its garbage collection action is not interfering with concurrent ongoing use.
This is supported and tested in the dataflow scheduler
by postponing execution of the last instruction in such a block.
This may cause a significant delay as a complete subquery could
be triggered.

A more elegant way is to inject sink() operations, which
are no-ops with side effects. Its argument list enumerates
all variables that can/should be released together.
[A side effect is that intermediates may hang around too long,
can be handled with a sink prepare instruction.TBD]