backing-out changeset 3669ddd28bf0 (on behalf of Martin)
Martin says:
"
I think I found it. In the new setup, thread creation/deletion
kills the performance on such loads.
I dropped the central workpool for a private one, but that
one was garbage collected at the end of a dataflow block.

simple insert in wisconsin could create/detroy 16 workers.

Have to find an alternative that retains the pool idea.
"

See also
http://mail.monetdb.org/pipermail/developers-list/2013-August/004002.html