Generate_series module checkin This directory contains a template to produce a single column table using generated data. (An extended version of the generate_series available in  Postgresql)
The program is optimized by on the fly replacement of the function vault.generate_series()
into generate_series.noop()  that allows to work on the parameters of the generator directly.
It shows how a (theta)select and leftfetchjoin function are replaced by a dedicated version
by self-modifying the MAL plan.

The current implementation does not exploit the candidate set arguments yet.

The collection could be extended to also accomodate arbitrary value joins against the generated
sequence, i.e. without its materialization first.