backout of changeset 9ff44367bf1d: it does not solve bug 3923, but rather limits a valuable debugging feature.
The symptoms reported in bug 3923 were/are not caused by bug in the code,
but rather by using an assertion-enabled (debug?-) build in a "production setting".
Simply disabling assertions avoid the symtopms reported in the bug report
(see also Sjoerd's comment in the bug report).

Changeset 9ff44367bf1d resulted in
(1) priting (one) warning regardless of whether assertions are enabled or not
(2) even with assertions enabled, only a single warning was issues,
    whichever happened to arbitrarily come first,
    while all subsequent (very likely different!) warnings were "swallowed"
    limiting the use of these warning severely if not removing it completely.

As a side not, the "scarry" part of bug 3923 is that the logs suggest
that we do not even have (or find) bulk operations for very common
standard opertions, including substration ("-") of 64-bit interger values
("lng") --- we surely should further investigate this,
and backing out changeset 9ff44367bf1d will help with this.