Fix assertion trigger on edge case where input BATs to UDF are empty.
This happens when there is a parallel function + filter, e.g.
SELECT parallel_function(i) FROM tbl WHERE attr > 50;

The partitioning happens before the filter function is executed, and thus certain partitions can be empty if the filter function is "unbalanced" (e.g. if you have values 1-100, and you filter on >50, the first few partitions will be empty while the others will be filled entirely).

This used to throw an error because we did not expect empty input BATs; but now we just return emptyBATs when this happens.