Remove these assertions again.
These assertions were removed for a reason. When a client connects using the new protocol the server switches from a bs_stream to a bs2_stream (so we can support compression and bigger blocks).

As part of this switch we want to close the old bs_stream but NOT the underlying stream (as we want to use that for our new bs2_stream).

These assertions were there to ensure that the underlying stream was also destroyed when the block stream was destroyed; however, we want to destroy ONLY the upper stream and not the underlying stream in this scenario.