Add support for long SQL texts in the profiler
This is related to Bug 6375, but is still incomplete.

The problem is that the profiler truncates JSON objects to 8KB. There are two
solutions to this problem:

  1. Since the location of the very big parts of the MAL plan are
  known (statement text, short statement text, and value in argument list) we
  can truncate only these parts, and thus produce a complete JSON object. The
  problem with this approach is that we do not know in advance how much we need
  to truncate, because other elements are to follow and everything needs to fit
  in 8KB.

  2. Transmit more than 8KB chunks. The obvious downside to this is that we will
  need more IO and since the query itself can in principle be unbounded, this
  can be very expensive.

The solution is the combination of the above. We truncate the big parts and so
effectively we make the amount of IO bounded but also recognizing that they
contain useful information, and try to minimize the information loss.

Still TODO:
The concatenation and transmission of big objects is buggy. The problem can be
reproduced with the command:

python -c "print(\"select * from _tables where name='\" + 'a'*8192 + \"';\")" | mclient -d